refactor(admob): share next-gen AdMob test fixtures
