+++ feature/admob-next-gen/src/test/kotlin/com/revenuecat/purchases/admob/nextgen/NativeAdFlowTest.kt
@@ -2,25 +2,17 @@
 
 package com.revenuecat.purchases.admob.nextgen
 
-import com.google.android.libraries.ads.mobile.sdk.banner.BannerAd
 import com.google.android.libraries.ads.mobile.sdk.banner.BannerAdEventCallback
-import com.google.android.libraries.ads.mobile.sdk.banner.BannerAdRefreshCallback
 import com.google.android.libraries.ads.mobile.sdk.common.LoadAdError
-import com.google.android.libraries.ads.mobile.sdk.common.ResponseInfo
-import com.google.android.libraries.ads.mobile.sdk.nativead.CustomNativeAd
-import com.google.android.libraries.ads.mobile.sdk.nativead.NativeAd
 import com.google.android.libraries.ads.mobile.sdk.nativead.NativeAdEventCallback
 import com.google.android.libraries.ads.mobile.sdk.nativead.NativeAdLoadResult
 import com.google.android.libraries.ads.mobile.sdk.nativead.NativeAdLoader
 import com.google.android.libraries.ads.mobile.sdk.nativead.NativeAdLoaderCallback
-import com.google.android.libraries.ads.mobile.sdk.nativead.NativeAdRequest
 import com.revenuecat.purchases.InternalRevenueCatAPI
-import com.revenuecat.purchases.Purchases
 import com.revenuecat.purchases.admob.nextgen.tracking.TrackingBannerAdEventCallback
 import com.revenuecat.purchases.admob.nextgen.tracking.TrackingBannerAdRefreshCallback
 import com.revenuecat.purchases.admob.nextgen.tracking.TrackingNativeAdEventCallback
 import com.revenuecat.purchases.ads.events.AdCaptureMethod
-import com.revenuecat.purchases.ads.events.AdTracker
 import com.revenuecat.purchases.ads.events.types.AdFailedToLoadData
 import com.revenuecat.purchases.ads.events.types.AdFormat
 import com.revenuecat.purchases.ads.events.types.AdLoadedData
@@ -40,32 +32,30 @@ import org.junit.Assert.assertNull
 import org.junit.Assert.assertSame
 import org.junit.Assert.assertTrue
 import org.junit.Before
+import org.junit.Rule
 import org.junit.Test
 class NativeAdFlowTest {
-    private val adTracker = mockk<AdTracker>(relaxed = true)
-    private val purchases = mockk<Purchases>(relaxed = true)
+    @get:Rule
+    val configuredPurchases = ConfiguredPurchasesRule()
+
+    private val adTracker get() = configuredPurchases.adTracker
 
     @Before
     fun setUp() {
-        every { purchases.adTracker } returns adTracker
         mockkObject(NativeAdLoader.Companion)
-        mockkObject(Purchases)
-        every { Purchases.isConfigured } returns true
-        every { Purchases.sharedInstance } returns purchases
     }
 
     @After
     fun tearDown() {
-        unmockkObject(Purchases)
         unmockkObject(NativeAdLoader.Companion)
     }
 
     @Test
     fun `callback load tracks and configures every possible success before forwarding`() {
-        val adRequest = adRequest("native-unit")
+        val adRequest = nativeAdRequest("native-unit")
         val nativeAd = nativeAd("native-network", "native-response")
         val customNativeAd = customNativeAd("custom-network", "custom-response")
-        val bannerAd = bannerAd("banner-network", "banner-response")
+        val bannerAd = nativeBannerAd("banner-network", "banner-response")
         val nativeEventCallback = mockk<NativeAdEventCallback>(relaxed = true)
         val bannerEventCallback = mockk<BannerAdEventCallback>(relaxed = true)
         val trackingLoadCallback = slot<NativeAdLoaderCallback>()
@@ -107,7 +97,7 @@ class NativeAdFlowTest {
 
     @Test
     fun `callback load tracks failure before forwarding`() {
-        val adRequest = adRequest("native-unit")
+        val adRequest = nativeAdRequest("native-unit")
         val error = mockk<LoadAdError> {
             every { code } returns LoadAdError.ErrorCode.NO_FILL
         }
@@ -131,7 +121,7 @@ class NativeAdFlowTest {
 
     @Test
     fun `suspending load returns and configures native result unchanged`() = runBlocking {
-        val adRequest = adRequest("native-unit")
+        val adRequest = nativeAdRequest("native-unit")
         val nativeAd = nativeAd("native-network", "native-response")
         val sdkResult = NativeAdLoadResult.NativeAdSuccess(nativeAd)
 
@@ -146,9 +136,9 @@ class NativeAdFlowTest {
 
     @Test
     fun `suspending load maps custom native banner and failure results`() = runBlocking {
-        val adRequest = adRequest("native-unit")
+        val adRequest = nativeAdRequest("native-unit")
         val customNativeAd = customNativeAd("custom-network", "custom-response")
-        val bannerAd = bannerAd("banner-network", "banner-response")
+        val bannerAd = nativeBannerAd("banner-network", "banner-response")
         val error = mockk<LoadAdError> {
             every { code } returns LoadAdError.ErrorCode.NETWORK_ERROR
         }
@@ -177,7 +167,7 @@ class NativeAdFlowTest {
     fun `tracking-safe setters preserve native custom native and banner wrappers`() {
         val nativeAd = nativeAd("native-network", "native-response")
         val customNativeAd = customNativeAd("custom-network", "custom-response")
-        val bannerAd = bannerAd("banner-network", "banner-response")
+        val bannerAd = nativeBannerAd("banner-network", "banner-response")
         nativeAd.installTrackingEventCallback(null, "feed", "native-unit")
         customNativeAd.installTrackingEventCallback(null, "feed", "native-unit")
         bannerAd.installTrackingCallbacks(null, null, "feed", "native-unit")
@@ -198,7 +188,7 @@ class NativeAdFlowTest {
     fun `tracking-safe setters directly assign callbacks when tracking is not installed`() {
         val nativeAd = nativeAd("native-network", "native-response")
         val customNativeAd = customNativeAd("custom-network", "custom-response")
-        val bannerAd = bannerAd("banner-network", "banner-response")
+        val bannerAd = nativeBannerAd("banner-network", "banner-response")
         nativeAd.adEventCallback = mockk(relaxed = true)
         customNativeAd.adEventCallback = mockk(relaxed = true)
         bannerAd.adEventCallback = mockk(relaxed = true)
@@ -215,65 +205,4 @@ class NativeAdFlowTest {
         assertSame(bannerCallback, bannerAd.adEventCallback)
     }
 
-    private fun adRequest(adUnitId: String): NativeAdRequest = mockk {
-        every { this@mockk.adUnitId } returns adUnitId
-    }
-
-    private fun nativeAd(network: String, responseId: String): NativeAd {
-        var callback: NativeAdEventCallback? = null
-        return mockk(relaxed = true) {
-            every { getResponseInfo() } returns responseInfo(network, responseId)
-            every { adEventCallback } answers { callback }
-            every { adEventCallback = any() } answers { callback = firstArg() }
-        }
-    }
-
-    private fun customNativeAd(network: String, responseId: String): CustomNativeAd {
-        var callback: NativeAdEventCallback? = null
-        return mockk(relaxed = true) {
-            every { getResponseInfo() } returns responseInfo(network, responseId)
-            every { adEventCallback } answers { callback }
-            every { adEventCallback = any() } answers { callback = firstArg() }
-        }
-    }
-
-    private fun bannerAd(network: String, responseId: String): BannerAd {
-        var callback: BannerAdEventCallback? = null
-        var refreshCallback: BannerAdRefreshCallback? = null
-        return mockk(relaxed = true) {
-            every { getResponseInfo() } returns responseInfo(network, responseId)
-            every { adEventCallback } answers { callback }
-            every { adEventCallback = any() } answers { callback = firstArg() }
-            every { bannerAdRefreshCallback } answers { refreshCallback }
-            every { bannerAdRefreshCallback = any() } answers { refreshCallback = firstArg() }
-        }
-    }
-
-    private fun responseInfo(network: String, responseId: String): ResponseInfo = mockk {
-        every { adapterClassName } returns network
-        every { this@mockk.responseId } returns responseId
-    }
-
-    private class RecordingNativeAdLoaderCallback : NativeAdLoaderCallback {
-        val nativeAds = mutableListOf<NativeAd>()
-        val customNativeAds = mutableListOf<CustomNativeAd>()
-        val bannerAds = mutableListOf<BannerAd>()
-        val loadErrors = mutableListOf<LoadAdError>()
-
-        override fun onNativeAdLoaded(nativeAd: NativeAd) {
-            nativeAds += nativeAd
-        }
-
-        override fun onCustomNativeAdLoaded(customNativeAd: CustomNativeAd) {
-            customNativeAds += customNativeAd
-        }
-
-        override fun onBannerAdLoaded(bannerAd: BannerAd) {
-            bannerAds += bannerAd
-        }
-
-        override fun onAdFailedToLoad(adError: LoadAdError) {
-            loadErrors += adError
-        }
-    }
 }
