+++ feature/admob-next-gen/src/test/kotlin/com/revenuecat/purchases/admob/nextgen/tracking/TrackingBannerAdRefreshCallbackTest.kt
@@ -4,9 +4,9 @@ package com.revenuecat.purchases.admob.nextgen.tracking
 
 import com.google.android.libraries.ads.mobile.sdk.banner.BannerAdRefreshCallback
 import com.google.android.libraries.ads.mobile.sdk.common.LoadAdError
-import com.google.android.libraries.ads.mobile.sdk.common.ResponseInfo
 import com.revenuecat.purchases.InternalRevenueCatAPI
 import com.revenuecat.purchases.Purchases
+import com.revenuecat.purchases.admob.nextgen.responseInfo
 import com.revenuecat.purchases.ads.events.AdCaptureMethod
 import com.revenuecat.purchases.ads.events.AdTracker
 import com.revenuecat.purchases.ads.events.types.AdFailedToLoadData
@@ -136,10 +136,4 @@ class TrackingBannerAdRefreshCallbackTest {
         verify(exactly = 0) { adTracker.trackAdLoaded(any(), any()) }
         assertTrue(delegated)
     }
-
-    private fun responseInfo(adapterClassName: String, responseId: String): ResponseInfo =
-        mockk<ResponseInfo>().also {
-            every { it.adapterClassName } returns adapterClassName
-            every { it.responseId } returns responseId
-        }
 }
